BLAN-1524 - refactoring of utilities, some utilities moved to workers for more convenience
