--- src/stores/sideBar.js
@@ -1,8 +1,0 @@
-import { defineStore } from 'pinia';
-
-export const useSideBarStore = defineStore('sideBarData', {
-  state: () => ({
-    isSideBarOpened: false,
-    isMobMenuActive: false
-  }),
-});
