+++ src/views/application/Index.vue
@@ -142,7 +142,6 @@ import { API } from '../../workers/api-worker/api.worker';
 import { VersionDetectorWorker } from '../../workers/version-detector-worker';
 import { useWindowWidth } from '../../workers/window-size-worker/widthScreen';
 import { useElementSize } from '@vueuse/core';
-import { validateRefreshToken } from '../../utils/validateRefreshToken';
 
 import EventUpdatedIcon from '../../assets/img/event-updated-modal-icon.svg';
 import EventCreatedIcon from '../../assets/img/event-creted-modal-icon.svg';
@@ -457,7 +456,6 @@ BlanballEventBus.on('closeScheduler', () => {
 });
 
 onMounted(() => {
-  validateRefreshToken();
   setHeaderHeight();
   setHeaderHeightCssVar();
 });
