+++ src/workers/api-worker/http/http-common/prefix.enum.authorization.js
@@ -10,5 +10,4 @@ export const AuthorizationEnpointsEnum = {
   ResetComplete: '/authentication/client/password/reset-complete',
   AddAvatar: '/authentication/client/update/my/profile/avatar',
   VerifyEmail: '/authentication/client/request-verify/email',
-  ValidateRefreshToken: '/authentication/client/validate/refresh/token',
 };
